configurable: simplified config syntax
Configuration of the separate components is now handled with a single
config_items class variable. This is a list of pairs specifying the
configuration variable name and if its required. The configuration
values could simply get via the __getitem__ method of the class
itself. Default values should be set in the constructor before the
constructor of Configurable.