Extract input/outputs
The graphios main now only creates multiple inputs/outputs and handles
the data exchange between them. So we removed the command line arguments
for specific inputs or outputs (carbon connections parameters and spool
path). All parser now needs a handler in their constructor.

The Config-class is now more simple because it does not need embedded
default values anymore. The config itself is more like a single list of
settings and two lists of input and outputs. The default values, that
are not at the top level (like the carbon connection parameters like
before) are handled by the special inputs or outputs itself.